shipyard_blueprints/googledrive/requirements.txt
google-api-python-client>=2.7
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0